+++ src/sql/cmd/profile.py
@@ -2,10 +2,24 @@ from sql import inspect
 from sql.cmd.cmd_utils import CmdParser
 
 
-def execute_profile_command(others):
+def profile(others):
     """
-    Execution logic for the profile command
+    Implementation of `%sqlcmd profile`
+    This function takes in a string containing command line arguments,
+    parses them to extract the name of the table, the schema, and the output location.
+    It then retrieves statistical information about the specified table and either
+    returns the report or writes it to the specified location.
 
+
+    Parameters
+    ----------
+    others : str,
+        A string containing the command line arguments.
+
+    Returns
+    -------
+    report: PrettyTable
+        statistics of the table
     """
     parser = CmdParser()
     parser.add_argument("-t", "--table", type=str, help="Table name", required=True)
